requirements.txt
Flask
gunicorn
requests
pyjwt
load_dotenv
pymsql
flask_sqlalchemy